requirements.txt
-e ./
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy
click
gevent
gunicorn
psycopg2
webargs==1.7.0
werkzeug
